nss: prevent PROTECT_ERRNO from squashing changes to *errnop
glibc passes in &errno for errnop, which means PROTECT_ERRNO ends up
squashing our intentional changes to *errnop.

This is a quick hack that adds a new flag alongside the saved value of
errno, which can be used to disarm PROTECT_ERRNO in the case that we
really do want to squash errno afterall.

Fixes #11321.